// cloudflare-worker/src/bindings.d.ts
export interface Env {
  SIGNER_CERT: string
  SIGNER_KEY: string
  SIGNER_KEY_PASSPHRASE: string
  PASS_TYPE_IDENTIFIER: string
  TEAM_IDENTIFIER: string
  GOOGLE_WALLET_SERVICE_ACCOUNT_EMAIL?: string
  GOOGLE_WALLET_SERVICE_ACCOUNT_PRIVATE_KEY?: string
  GOOGLE_WALLET_CLASS_ID?: string
  GOOGLE_WALLET_ALLOWED_ORIGINS?: string
  GOOGLE_WALLET_FRONTEND_URL?: string
  RUNPASS_TEST_PKPASS?: string
}